+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.4.0 - 2017-01-06
+
+- Change: The parameter `proget.apiKeyMapping.server` is back to be a `string` because "RegExp in configuration file is ugly. A lot more if you need to double escape it." (@franknarf8 comment :wink:).
+- Add: Tool configuration retro-compatibility support!
+
 ## v0.3.0 - 2016-12-19
 
 - Add: Now some warning will be displayed if you still have some "old proget-universal-bower-resolver configuration" parameters in your `.bowerrc` config file.